cnn models/requirements.txt
flask
tensorflow
pillow
numpy
flask-cors
Pillow